#15 Refactoring in TorrentChecker / TorrentUploader - created the TorrentDiskLoader class, which takes care of matching blocks and pieces to data written to disk
